requirements.txt
adr-tools-python==1.0.3
arrow==1.3.0
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
diagrams==0.23.4
graphviz==0.20.1
idna==3.6
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
Pygments==2.17.2
python-dateutil==2.8.2
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
text-unidecode==1.3
typed_ast==1.5.5
types-python-dateutil==2.8.19.20240106
urllib3==2.2.0
clipboard
gitpython
pathlib